client-hints/accept_ch_lifetime_cross_origin_iframe.tentative.sub.https.html

<html>
<title>Accept-CH-Lifetime test with cross-origin iframe</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=test>
    <p>Apart from this webpage, the test opens another html web page. One test
    is run in this web page, and another in the second web page.
</div>

<script>

// This test fetches resources/accept_ch_lifetime.html in a cross origin iframe.
// The response headers to that webpage contain Accept-CH and Accept-CH-Lifetime
// headers.

// Fetching that webpage should NOT cause the user-agent to persist origin
// preferences for the client hints specified.

// Next, to verify if the origin preferences were NOT persisted by the user
// agent, this test fetches resources/do_not_expect_client_hints_headers.html
// in a new window. Fetching of
// resources/do_not_expect_client_hints_headers.html
// verifies that the user agent did not actually sent the client hints in the
// request headers.

// Test is marked as tentative until https://github.com/whatwg/fetch/issues/726
// is resolved.

// First, verify the initial state to make sure that the browser does not have
// client hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Precondition: Test that the browser does not have client hints preferences cached");

</script>

<!-- Fetching this webpage should NOT cause user-agent to persist client hint
 preferences for the origin.-->
<iframe src="https://{{domains[www]}}:{{ports[https][0]}}/client-hints/resources/accept_ch_lifetime.html"></iframe>

<script>
// Open a new window. Verify that the user agent does not attach the client
// hints.
window.open("resources/do_not_expect_client_hints_headers.html");
async_test(t => {
window.addEventListener('message', function(event) {
  t.done();
})
}, "Loading of resources/do_not_expect_client_hints_headers.html did not finish.");


</script>
</body>
</html>
